Change doker_id with container_id and and introduce multi-tenancy switch off
